Audio
Tried to implement the USynthComponent System. I searched online for C++ implementation. Only found BP one. But I tried to do it via script and ended up with linker errors. Same for the collision.
As for the UI, there is an error that comes up due to something that I did wrong.